@@ -12,33 +12,63 @@
     "start": "set APP_PORT=3000&& next start -p 3000",
     "start80": "set APP_PORT=80&& next start -p 80",
     "start80p": "set APP_PORT=80&&set URL_DESTROYER=asdf&& next start -p 80",
-    "lint": "next lint"
+    "lint": "next lint",
+    "prisma_seed_all": "npx prisma db seed",
+    "prisma_studio": "npx prisma studio",
+    "prisma_db_push": "npx prisma db push",
+    "prisma_generate": "npx prisma generate",
+    "jest-test": "npm test",
+    "playwright-test": "npx playwright test",
+    "test": "jest",
+    "test:watch": "jest --watch"
+  },
+  "prisma": {
+    "seed": "ts-node --compiler-options {\"module\":\"CommonJS\"} prisma/seeds/index.ts"
   },
   "dependencies": {
     "@next/bundle-analyzer": "^12.0.10",
-    "@reduxjs/toolkit": "^1.7.2",
+    "@prisma/client": "^6.6.0",
+    "@reduxjs/toolkit": "^2.7.0",
+    "@types/jsonwebtoken": "^9.0.9",
+    "@types/react": "19.1.2",
+    "@types/react-dom": "19.1.3",
+    "bcryptjs": "^3.0.2",
     "better-sqlite3": "^8.1.0",
     "bootstrap": "^5.1.3",
     "clone": "^2.1.2",
     "dompurify": "^2.3.6",
     "formidable": "^1.2.6",
     "iron-session": "^8.0.4",
+    "jsonwebtoken": "^9.0.2",
     "madge": "^5.0.1",
-    "next": "^13.5.5",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-redux": "^7.2.6",
+    "next": "^15.3.0",
+    "next-auth": "^4.24.11",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "react-redux": "^9.2.0",
     "redux-saga": "^1.1.3",
+    "server-only": "^0.0.1",
     "sharp": "^0.32.6"
   },
   "devDependencies": {
-    "@types/node": "^17.0.17",
-    "@types/react": "^17.0.39",
+    "@playwright/test": "^1.52.0",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^17.0.45",
     "cross-env": "^7.0.3",
     "eslint": "7.30.0",
     "eslint-config-next": "^13.5.5",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^30.0.0-beta.3",
+    "playwright": "^1.52.0",
+    "prisma": "^6.6.0",
     "sass": "^1.49.7",
-    "typescript": "^4.5.5",
+    "ts-jest": "^29.3.4",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.8.3",
     "webp-loader": "^0.6.0"
   }
 }
