@@ -1,4 +1,4 @@
-// eslint-disable-next-line
+"use client"
 
 import React, { useEffect, useState } from "react";
 import { DetailFrameMode } from "../../../../FilesToDistribute/constants";
@@ -9,17 +9,20 @@ import { useSelector } from "react-redux";
 import { setErrorMsg } from "../../../../store/reducers/ErrorSlice";
 import { setDBObject, setEditedAttrs, setNewDBObject } from "../../../../store/reducers/DBObjectSlice";
 import { setEntries } from "../../../../store/reducers/EntrySlice";
-import { getRawDBObjectDefinition } from "../../../data/definitions/db-object-definitions";
-import useAppDispatch from "../../../../shared/hooks/useAppDispatch";
+import useAppDispatch from "../../../../hooks/useAppDispatch";
 import { selectActualFormDefinition } from "../../../../store/formDefReducer/selector";
+import { SagaActions } from "@store/sagaActions";
 
 type FormFrameContainerProps = {
+    formID: string
 
 }
 
 const FormFrameContainer: React.FC<FormFrameContainerProps> = ({
-
+    formID
 }) => {
+
+
     const dispatch = useAppDispatch();
     const definition = useSelector((state: RootState) => selectActualFormDefinition(state));
     let DBOClass = definition?.DB?.DBOClass ?? undefined;
@@ -32,28 +35,41 @@ const FormFrameContainer: React.FC<FormFrameContainerProps> = ({
     const [saveDialogVisible, setSaveDialogVisible] = useState(false);
     const [detailFrameMode, setDetailFrameMode] = useState(DetailFrameMode.NEW_ENTRY);
     const [detailFrameVisible, setDetailFrameVisible] = useState(false)
-    const entries = useSelector((state: RootState) => state.entries);
 
-    const DBObject = useSelector((state: RootState) => state.dbObject);
+    const DBObject: DBObjectType = useSelector((state: RootState) => state.dbObject);
 
 
-    useEffect(() => {        
-        let detailItemCondition = "";
+
+
+
+    const [pageFormDefinitionLoaded, setPageFormDefinitionLoaded] = useState(false);
+
+    useEffect(() => {
+        dispatch({ type: SagaActions.SET_FORM_DEFINITIONS, FID: formID });
+    }, [dispatch])
+
+    useEffect(() => {
+        setPageFormDefinitionLoaded(true)
+    }, [DBOClass])
+
+
+
+
+    useEffect(() => {
+        let detailItemCondition;
         let parentAttribute = DBManager.getEmptyDBObject(DBOClass)?.persistentAttributes[0];
         if (breadcrumbItems.length) {
 
             if (parentAttribute) {
                 let key: string = parentAttribute.key;
                 key = (key.startsWith("*")) ? key.substring(1) : key;
-                detailItemCondition = `WHERE ${key}='${DBManager.getAttrFromArrByKey(breadcrumbItems[breadcrumbItems.length - 1].DBObject.attributes, parentAttribute.key).value}'`;
+                detailItemCondition = JSON.stringify({[key]: DBManager.getAttrFromArrByKey(breadcrumbItems[breadcrumbItems.length - 1].DBObject.attributes, parentAttribute.key).value});
             }
         }
-        DBManager.getAllDBObjectEntries(DBOClass, definition?.DB?.orderBy,detailItemCondition).then(entrs => {
+        DBManager.getAllDBObjectEntries(DBOClass, definition?.DB?.orderBy, detailItemCondition).then(entrs => {
             dispatch(setEntries(entrs));
-            let pa = getRawDBObjectDefinition(DBOClass)?.persistentAttributes ?? [];
-            //dispatch(setPersistentAttrs(pa))
         })
-        if(DBObject.DBOClass == undefined && DBOClass !== undefined && DBOClass.length){
+        if (DBObject.DBOClass == undefined && DBOClass !== undefined && DBOClass.length) {
             dispatch(setNewDBObject({ DBOClass, parentEntry: undefined }));
         }
         // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -61,9 +77,6 @@ const FormFrameContainer: React.FC<FormFrameContainerProps> = ({
 
 
     const showDetailFrame = () => {
-        /*if (setEmptyObject) {
-          setDBObject(getEmptyDBObject(formClassName[formClassName.length - 1]));
-        }*/
         setDetailFrameVisible(true);
     }
 
@@ -91,7 +104,9 @@ const FormFrameContainer: React.FC<FormFrameContainerProps> = ({
 
     return (
         <>
-            <FormFrame errorMsg={errorMsg} detailFrameVisible={detailFrameVisible} saveDialogVisible={saveDialogVisible} entries={entries} detailFrameMode={detailFrameMode} definition={definition} DBObject={DBObject} hideDetailFrame={hideDetailFrame} setDBObject={setDBObject} editItemHandler={editItemHandler} showDetailFrame={showDetailFrame} setSaveDialogVisible={setSaveDialogVisible} setErrorMsg={setErrMsg} createEntryText={definition.detailFrame.createNewEntryText} />
+            {
+            pageFormDefinitionLoaded && <FormFrame errorMsg={errorMsg} detailFrameVisible={detailFrameVisible} saveDialogVisible={saveDialogVisible} detailFrameMode={detailFrameMode} hideDetailFrame={hideDetailFrame} editItemHandler={editItemHandler} showDetailFrame={showDetailFrame} setSaveDialogVisible={setSaveDialogVisible} setErrorMsg={setErrMsg} createEntryText={definition.detailFrame.createNewEntryText} />
+            }
         </>
     )
 }
